requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-paypal==2.0
idna==3.4
Pillow==10.1.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.7
